requirements.txt
matplotlib==3.10.5
prettytable==3.16.0
pymongo==4.13.2
PyMySQL==1.1.1
python-dotenv==1.1.1
